pyproject.toml
[tool.poetry]
name = "flatiron-snowflake-loader"
version = "0.1.0"
description = ""
authors = ["Sean Davis <seandavi@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "flatiron_snowflake_loader"}]

[tool.poetry.dependencies]
python = "^3.9"
snowflake-connector-python = "^3.0.4"
pydantic = "^1.10.9"
click = "^8.1.3"
loguru = "^0.7.0"

[tool.poetry.scripts]
flatiron_loader = 'flatiron_snowflake_loader.loader:process'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
